Exception Handling V2

+++ src/main/java/com/ensa/msproduct/services/ProductServiceImpl.java
@@ -2,6 +2,9 @@ package com.ensa.msproduct.services;
 
 import com.ensa.msproduct.dao.ProductRepository;
 import com.ensa.msproduct.entities.Product;
+import com.ensa.msproduct.exceptions.EntityAlreadyExistsException;
+import com.ensa.msproduct.exceptions.EntityNotFoundException;
+import com.ensa.msproduct.exceptions.NegativeValuesException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,17 +26,29 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public Product getProductByDesignation(String designation) {
+        if(productRepository.findProductByDesignation(designation) == null ){
+            throw new EntityNotFoundException("The product doesn't exists !");
+        }
         return productRepository.findProductByDesignation(designation);
     }
 
     @Override
     public Product addProduct(Product product) {
+        if(productRepository.findProductById(product.getId()) != null ){
+            throw new EntityAlreadyExistsException("The product already exists");
+        }
+        if(product.getPrice() < 0 || product.getDepositQuantity() < 0){
+            throw new NegativeValuesException("The product's price or quantity cannot be negative");
+        }
         return productRepository.save(product);
     }
 
     @Override
     public Product editProduct(Long id, Product product) {
 
+        if(productRepository.findProductById(id) == null ){
+            throw new EntityNotFoundException("The product doesn't exists");
+        }
         Product updatedProduct = productRepository.findById(id).get();
         updatedProduct.setDesignation(product.getDesignation());
         updatedProduct.setPrice(product.getPrice());
@@ -41,17 +56,18 @@ public class ProductServiceImpl implements ProductService{
         updatedProduct.setDepositQuantity(product.getDepositQuantity());
         updatedProduct.setExpirationDate(product.getExpirationDate());
         updatedProduct.setShortDescription(product.getShortDescription());
+        if(product.getPrice() < 0 || product.getDepositQuantity() < 0){
+            throw new NegativeValuesException("The product's price or quantity cannot be negative");
+        }
 
         return productRepository.save(updatedProduct);
     }
 
     @Override
-    public Product findProductById(Long id) {
-        return productRepository.findProductById(id);
-    }
-
-    @Override
     public void deleteProduct(Long id) {
+        if(productRepository.findProductById(id) == null ){
+            throw new EntityNotFoundException("The product doesn't exists");
+        }
         productRepository.deleteById(id);
     }
 }
